In the Visual Studio libLouis.h, widechar is now an int rather than short. This allows properly working with  ancient language tables. Because of include precedence, we remove any generated libLouis.h in the libLouis directory to ensure that the one in the VisualStudio directory is used instead.
We also have a Visual Studio specific strings.h that  just has the line
defines strncasecmp to be strnicmp

+++ VisualStudio/liblouis.h
@@ -37,7 +37,10 @@ extern "C" {
 
 // clang-format interacts badly with @foo@
 // clang-format off
-typedef unsigned short int widechar;
+	// widechar needs to be unsigned int in order to support ancient
+	// languages since their code points are in a supplemental Unicode
+	// plane with values above 0xffff.
+typedef unsigned int widechar;
 // clang-format on
 typedef unsigned short formtype;
